requirements.txt
Flask
google-generativeai
flask-cors
python-dotenv
